// src/shared/api/generated/model/tourSpot.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 아이랑 - 홈 화면 API
 * OpenAPI spec version: 0.0.1
 */
import type { Region } from './region';

export interface TourSpot {
  name: string;
  id: string;
  region: Region;
  /** @nullable */
  imgUrl: string | null;
  isStrollerRental: boolean;
  isPark: boolean;
  isToilet: boolean;
  /** @nullable */
  operatingHours: string | null;
  /** @nullable */
  price: string | null;
  latitude: number;
  longitude: number;
}